test(refactor): extract early-crash code.
